# openapi.yaml
openapi: 3.0.3
info:
  title: BarberPOS Backend API
  version: 0.1.0
servers:
  - url: https://api.barberpos.cloud
paths:
...
  /categories:
    get:
      summary: List categories
      security:
        - bearerAuth: []
      responses:
        '200':
          description: List
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponse'
                  - type: object
                    properties:
                      data:
                        type: array
                        items:
                          $ref: '#/components/schemas/Category'
    post:
      summary: Upsert category
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                id:
                  type: integer
                  format: int64
                name:
                  type: string
      responses:
        '200':
          description: Saved
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponse'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/Category'
  /categories/{id}:
    delete:
      summary: Delete category
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: Deleted
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponse'
                  - type: object
                    properties:
                      data:
                        type: object
                        properties:
                          ok: { type: boolean }
  /customers:
    get:
      summary: List customers
      security:
        - bearerAuth: []
      responses:
        '200':
          description: List
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponse'
                  - type: object
                    properties:
                      data:
                        type: array
                        items:
                          $ref: '#/components/schemas/Customer'
    post:
      summary: Upsert customer
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Customer'
      responses:
        '200':
          description: Saved
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponse'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/Customer'
  /customers/{id}:
    delete:
      summary: Delete customer
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: Deleted
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponse'
                  - type: object
                    properties:
                      data:
                        type: object
                        properties:
                          ok: { type: boolean }
...
  /closing/summary:
    get:
      summary: Closing summary (today)
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Summary
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponse'
                  - type: object
                    properties:
                      data:
                        type: object
                        properties:
                          totalCash: { type: integer }
                          totalNonCash: { type: integer }
                          totalCard: { type: integer }
  /closing:
    get:
      summary: List closing history
      security:
        - bearerAuth: []
      parameters:
        - in: query
          name: limit
          schema:
            type: integer
            default: 50
      responses:
        '200':
          description: List
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponse'
                  - type: object
                    properties:
                      data:
                        type: array
                        items:
                          $ref: '#/components/schemas/ClosingHistory'
    post:
      summary: Submit closing
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                tanggal:
                  type: string
                  example: 2025-01-01
                shift:
                  type: string
                karyawan:
                  type: string
                shiftId:
                  type: string
                operatorName:
                  type: string
                total:
                  type: integer
                status:
                  type: string
                catatan:
                  type: string
                fisik:
                  type: string
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponse'
                  - type: object
                    properties:
                      data:
                        type: object
                        properties:
                          ok: { type: boolean }
                          id: { type: integer, format: int64 }
...
  /employees:
    get:
      summary: List employees
      security:
        - bearerAuth: []
      responses:
        '200':
          description: List
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponse'
                  - type: object
                    properties:
                      data:
                        type: array
                        items:
                          type: object
                          properties:
                            id: { type: integer, format: int64 }
                            managerUserId: { type: integer, format: int64, nullable: true }
                            name: { type: string }
                            role: { type: string }
                            phone: { type: string }
                            email: { type: string }
                            joinDate: { type: string }
                            commission: { type: number, nullable: true }
                            active: { type: boolean }
    post:
      summary: Upsert employee
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                id: { type: integer }
                name: { type: string }
                role: { type: string }
                phone: { type: string }
                email: { type: string }
                pin: { type: string, description: "Set/replace PIN (store as hash)" }
                joinDate: { type: string, example: 2025-01-01 }
                commission: { type: number, format: float }
                active: { type: boolean }
      responses:
        '200':
          description: Saved
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponse'
                  - type: object
                    properties:
                      data:
                        type: object
                        properties:
                          id: { type: integer, format: int64 }
                          managerUserId: { type: integer, format: int64, nullable: true }
                          name: { type: string }
                          role: { type: string }
                          phone: { type: string }
                          email: { type: string }
                          joinDate: { type: string }
                          commission: { type: number, nullable: true }
                          active: { type: boolean }
  /employees/{id}:
    delete:
      summary: Delete employee
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Deleted
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponse'
                  - type: object
                    properties:
                      data:
                        type: object
                        properties:
                          ok: { type: boolean }
  /notifications/token:
    post:
      summary: Register FCM token
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                token:
                  type: string
                platform:
                  type: string
      responses:
        '200':
          description: Saved
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponse'
                  - type: object
                    properties:
                      data:
                        type: object
                        properties:
                          ok: { type: boolean }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    ApiError:
      type: object
      properties:
        code: { type: integer }
        status: { type: string }
    ApiResponse:
      type: object
      required: [status, data]
      properties:
        status: { type: string, enum: [success, error] }
        message: { type: string }
        data: {}
        error:
          $ref: '#/components/schemas/ApiError'
...
    Category:
      type: object
      properties:
        id: { type: integer, format: int64 }
        name: { type: string }
    Customer:
      type: object
      properties:
        id: { type: integer, format: int64 }
        name: { type: string }
        phone: { type: string }
        email: { type: string }
        address: { type: string }
...
    ClosingHistory:
      type: object
      properties:
        id: { type: integer, format: int64 }
        tanggal: { type: string }
        shift: { type: string }
        karyawan: { type: string }
        shiftId: { type: string }
        operatorName: { type: string }
        total: { type: integer }
        status: { type: string }
        catatan: { type: string }
        fisik: { type: string }
        createdAt: { type: string, format: date-time }
...